refactor(cli): extract shared output module, migrate jq runner
Move the CLI output helpers duplicated between the jq and yq runners
into a new src/bin/succinctly/output.rs, following the env_config
precedent: exit_codes (full jq set incl. HALT_ERROR), JSON string
escapers, ColorScheme/JQ_COLORS parsing, the token-aware colorizer,
print_build_configuration (parameterized by tool name), and the
recursive JSON pretty-printer.

The pretty-printer is jq's format_json_impl verbatim, plus two knobs
so both runners can share it: sort_keys now sorts object entries
inline (replacing the clone-based sort_keys() pre-pass and
format_json_sorted), and FloatStyle selects between jq's shortest
float form and yq's trailing-.0 preservation. jq output is
byte-identical; verified against jq_cli_tests, cli_golden_tests, and
cli_characterization_tests.

Unit tests for the moved helpers move to output.rs, with a new
regression test pinning token-aware colorization.

Refs #181
